+++ staticpipes/config.py
@@ -1,9 +1,14 @@
 class Config:
 
-    def __init__(self, pipes: list = [], context: dict = {}, checks: list = []):
+    def __init__(
+        self,
+        pipes_and_groups_of_pipes: list | None = None,
+        context: dict = {},
+        checks: list = [],
+    ):
         # Pipes
-        self.pipes: list = pipes
-        for pipe in self.pipes:
+        self._pipes_and_groups_of_pipes: list = pipes_and_groups_of_pipes or []
+        for pipe in self._pipes_and_groups_of_pipes:
             pipe.config = self
         # Context
         self.context: dict = context
@@ -12,5 +17,5 @@ class Config:
         for check in checks:
             check.config = self
 
-    def get_pipes(self) -> list:
-        return self.pipes
+    def get_pipes_and_groups_of_pipes(self) -> list:
+        return self._pipes_and_groups_of_pipes