+++ tests/test_collection_csv.py
@@ -14,7 +14,7 @@ def test_collection_csv():
     # setup
     out_dir = tempfile.mkdtemp(prefix="staticpipes_tests_")
     config = staticpipes.config.Config(
-        pipes=[
+        pipes_and_groups_of_pipes=[
             staticpipes.pipes.load_collection_csv.PipeLoadCollectionCSV(
                 filename="data.csv"
             ),
@@ -63,7 +63,7 @@ def test_collection_csv_dir_mode():
     # setup
     out_dir = tempfile.mkdtemp(prefix="staticpipes_tests_")
     config = staticpipes.config.Config(
-        pipes=[
+        pipes_and_groups_of_pipes=[
             staticpipes.pipes.load_collection_csv.PipeLoadCollectionCSV(
                 filename="data.csv"
             ),
@@ -103,7 +103,7 @@ def test_pipe_collection_records_process_collection_with_filter_function():
     # setup
     out_dir = tempfile.mkdtemp(prefix="staticpipes_tests_")
     config = staticpipes.config.Config(
-        pipes=[
+        pipes_and_groups_of_pipes=[
             staticpipes.pipes.load_collection_csv.PipeLoadCollectionCSV(
                 filename="data.csv"
             ),